Feature engineered RF classifier based on Gustaf's code
I worked a bit on the original Random Forest code uploaded by Gustaf. In case any one of you is keen and would have some time, it seems to me that passing on the data to the classifier after "taking out" Gear 0 using a rule could be one way to enhance further the model, which already works quite well. This can result in remarkable performance on the testing data, but the confusion matrices won't contain any information on the data which was initially classified as Gear 0, the accuracy of that has to be assessed separately to evaluate the overall performance. In addition, I worked on the features passed on to the model. I don't think it is a must for you to work on the models anymore, the priority should be the writing process, I just shared this in case some of you would like to work on it as an "extra", or if you would consider reporting its findings among your results.